pyproject.toml
[tool.poetry]
name = "chatformula1"
version = "0.1.0"
description = "AI-powered Formula 1 expert chatbot with RAG architecture"
authors = ["Prateek Mulye <prateekmulye@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"

# Core Dependencies
python-dotenv = "^1.0.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"

# LangChain & LangGraph - Latest Stable Versions from PyPI
langchain = "^1.1.0"
langgraph = "^1.0.4"

# LangChain Integrations - Latest Stable
langchain-openai = "^1.1.0"
langchain-pinecone = "^0.2.13"
langchain-tavily = "^0.2.13"
langchain-community = "^0.4.1"
langchain-text-splitters = "^1.0.0"

# Vector Store Dependencies - Latest with async support
pinecone = {extras = ["asyncio"], version = "^7.3.0"}

# Additional LangChain Tools - Compatible with LangChain 1.1.0+
langsmith = "^0.3.45"

# Logging - Latest
structlog = "^24.4.0"
python-json-logger = "^3.3.0"

# API Framework - Latest
fastapi = "^0.115.14"
uvicorn = {extras = ["standard"], version = "^0.34.3"}

# UI Framework - Latest
streamlit = "^1.51.0"

# Utilities - Latest
httpx = "^0.28.1"
tenacity = "^9.0.0"
aiofiles = "^24.1.0"
click = "^8.1.8"

[tool.poetry.group.dev.dependencies]
# Testing - Latest
pytest = "^8.4.2"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.3.0"
pytest-mock = "^3.15.1"

# Code Quality - Latest
black = "^25.11.0"
ruff = "^0.8.6"
mypy = "^1.19.0"
pre-commit = "^4.5.0"

# Type Stubs - Latest
types-aiofiles = "^24.1.0.20250822"

[tool.poetry.scripts]
f1-api = "src.api.main:main"
f1-ui = "src.ui.app:main"
f1-ingest = "src.ingestion.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "S",   # flake8-bandit
    "T20", # flake8-print
]
ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ANN101", # missing type annotation for self
    "ANN102", # missing type annotation for cls
    "S101",   # use of assert
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "ANN"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langgraph.*",
    "streamlit.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
